+++ packages/imperative/__tests__/__integration__/cmd/__tests__/integration/cli/auth/Cmd.cli.auth.logout.fruit.integration.test.ts
@@ -12,11 +12,10 @@
 import { runCliScript } from "../../../../../../src/TestUtil";
 import { ITestEnvironment } from "../../../../../../__src__/environment/doc/response/ITestEnvironment";
 import { SetupTestEnvironment } from "../../../../../../__src__/environment/SetupTestEnvironment";
-import { join } from "path";
 
 // Test Environment populated in the beforeAll();
 let TEST_ENVIRONMENT: ITestEnvironment;
-describe("cmd-cli auth logout", () => {
+describe("imperative-test-cli auth logout", () => {
     // Create the unique test environment
     beforeAll(async () => {
         TEST_ENVIRONMENT = await SetupTestEnvironment.createTestEnv({
@@ -25,69 +24,71 @@ describe("cmd-cli auth logout", () => {
         });
     });
 
-    afterEach(() => {
-        // delete profiles between tests so that they can be recreated
-        require("rimraf").sync(join(TEST_ENVIRONMENT.workingDir, "profiles"));
-    });
-
-    it("should have auth logout command that loads values from base profile and removes the token with alias", () => {
-        let response = runCliScript(__dirname + "/__scripts__/base_profile_and_auth_li.sh",
-            TEST_ENVIRONMENT.workingDir, ["fakeUser", "fakePass"]);
-        expect(response.stderr.toString()).toContain("command 'profiles create' is deprecated");
-        expect(response.status).toBe(0);
-
-        // the output of the command should include token value
-        expect(response.stdout.toString()).toContain("tokenType:  jwtToken");
-        expect(response.stdout.toString()).toContain("tokenValue: fakeUser:fakePass@fakeToken");
-
-        response = runCliScript(__dirname + "/__scripts__/base_profile_and_auth_lo.sh",
+    it("should have auth lo command that loads values from base profile and removes the token", () => {
+        let response = runCliScript(__dirname + "/__scripts__/auth_li_config_password.sh",
             TEST_ENVIRONMENT.workingDir);
-        expect(response.stderr.toString()).toContain("command 'profiles list' is deprecated");
+
+        // the output of the login command should include token value
+        expect(response.stderr.toString()).toBe("");
+        expect(response.stdout.toString()).toContain("tokenType:  jwtToken");
+        expect(response.stdout.toString()).toContain("tokenValue: (secure value)");
         expect(response.status).toBe(0);
 
-        // the output of the command should include token value
+        response = runCliScript(__dirname + "/__scripts__/auth_lo.sh",
+            TEST_ENVIRONMENT.workingDir);
+
+        // the output of the command should NOT include token value
+        expect(response.stderr.toString()).toBe("");
+        expect(response.stdout.toString()).toContain("Logout successful. The authentication token has been revoked");
+        expect(response.stdout.toString()).toContain("Token was removed from your 'baseProfName_fruit' base profile");
         expect(response.stdout.toString()).not.toContain("tokenType:");
         expect(response.stdout.toString()).not.toContain("tokenValue:");
+        expect(response.status).toBe(0);
     });
 
     it("should have auth logout command that loads values from base profile and removes the token", () => {
-        let response = runCliScript(__dirname + "/__scripts__/base_profile_and_auth_login.sh",
-            TEST_ENVIRONMENT.workingDir, ["fakeUser", "fakePass"]);
-        expect(response.stderr.toString()).toContain("command 'profiles create' is deprecated");
-        expect(response.status).toBe(0);
-
-        // the output of the command should include token value
-        expect(response.stdout.toString()).toContain("tokenType:  jwtToken");
-        expect(response.stdout.toString()).toContain("tokenValue: fakeUser:fakePass@fakeToken");
-
-        response = runCliScript(__dirname + "/__scripts__/base_profile_and_auth_logout.sh",
+        let response = runCliScript(__dirname + "/__scripts__/auth_login_config_password.sh",
             TEST_ENVIRONMENT.workingDir);
-        expect(response.stderr.toString()).toContain("command 'profiles list' is deprecated");
+
+        // the output of the login command should include token value
+        expect(response.stderr.toString()).toBe("");
+        expect(response.stdout.toString()).toContain("tokenType:  jwtToken");
+        expect(response.stdout.toString()).toContain("tokenValue: (secure value)");
         expect(response.status).toBe(0);
 
-        // the output of the command should include token value
+        response = runCliScript(__dirname + "/__scripts__/auth_logout.sh",
+            TEST_ENVIRONMENT.workingDir);
+
+        // the output of the command should NOT include token value
+        expect(response.stderr.toString()).toBe("");
+        expect(response.stdout.toString()).toContain("Logout successful. The authentication token has been revoked");
+        expect(response.stdout.toString()).toContain("Token was removed from your 'baseProfName_fruit' base profile");
         expect(response.stdout.toString()).not.toContain("tokenType:");
         expect(response.stdout.toString()).not.toContain("tokenValue:");
+        expect(response.status).toBe(0);
     });
 
     it("should have auth logout command that invalidates another token", () => {
-        let response = runCliScript(__dirname + "/__scripts__/base_profile_and_auth_login.sh",
-            TEST_ENVIRONMENT.workingDir, ["fakeUser", "fakePass"]);
-        expect(response.stderr.toString()).toContain("command 'profiles create' is deprecated");
-        expect(response.stderr.toString()).toContain("command 'profiles list' is deprecated");
+        let response = runCliScript(__dirname + "/__scripts__/auth_login_config_password.sh",
+            TEST_ENVIRONMENT.workingDir);
+
+        // the output of the login command should include token value
+        expect(response.stderr.toString()).toBe("");
+        expect(response.stdout.toString()).toContain("tokenType:  jwtToken");
+        expect(response.stdout.toString()).toContain("tokenValue: (secure value)");
         expect(response.status).toBe(0);
 
-        // the output of the command should include token value
-        expect(response.stdout.toString()).toContain("tokenType:  jwtToken");
-        expect(response.stdout.toString()).toContain("tokenValue: fakeUser:fakePass@fakeToken");
 
-        response = runCliScript(__dirname + "/__scripts__/base_profile_and_auth_logout_specify_token.sh",
+        response = runCliScript(__dirname + "/__scripts__/auth_logout_specify_token.sh",
             TEST_ENVIRONMENT.workingDir, ["fakeToken:fakeToken@fakeToken"]);
-        expect(response.stderr.toString()).toContain("command 'profiles list' is deprecated");
-        expect(response.status).toBe(0);
 
-        // the output of the command should include token value
+        // the output of the command should still include token value
+        expect(response.stderr.toString()).toBe("");
+        expect(response.stdout.toString()).toContain("Logout successful. The authentication token has been revoked");
+        expect(response.stdout.toString()).toContain("Token was not removed from your 'baseProfName_fruit' base profile");
+        expect(response.stdout.toString()).toContain("Reason: Token value does not match the securely stored value");
         expect(response.stdout.toString()).toContain("tokenType:  jwtToken");
-        expect(response.stdout.toString()).toContain("tokenValue: fakeUser:fakePass@fakeToken");
+        expect(response.stdout.toString()).toContain("tokenValue: (secure value)");
+        expect(response.status).toBe(0);
     });
 });