+++ packages/imperative/__tests__/__integration__/cmd/__tests__/integration/cli/auth/__scripts__/auth_logout.sh
@@ -1,0 +1,12 @@
+#!/bin/sh
+
+# include exitOnFailure function
+myScriptDir=`dirname $0`
+. $myScriptDir/exitOnFailure.sh
+
+imperative-test-cli auth logout fruit
+exitOnFailure "Logging out auth of type fruit failed!" $?
+
+# show contents of our config file
+imperative-test-cli config list profiles
+exitOnFailure "Display of updated imperative-test-cli.config.json failed!" $?