+++ packages/imperative/__tests__/__integration__/cmd/__tests__/integration/cli/profiles/__scripts__/profiles/map_banana_to_options.sh
@@ -1,21 +1,23 @@
-#!/bin/bash
+#!/bin/sh
 
-color=$1
-description=$2
-moldtype=$3
-# First create a banana profile
-cmd-cli profiles create banana-profile "test_banana" --color "$color" --banana-description "$description" --mold-type "$moldtype"
-CMDRC=$?
-if [ $CMDRC -gt 0 ]
-then
-    echo "Creating a test_banana profile of type banana failed!" 1>&2
-    exit $CMDRC
-fi
+profileColor=$1
+profileDescription=$2
+profileMoldType=$3
 
+# include exitOnFailure function
+myScriptDir=`dirname $0`
+. $myScriptDir/exitOnFailure.sh
+
+# set desired properties in our config file
+cp $myScriptDir/banana.config.json .
+exitOnFailure "Failed to copy config file." $?
+
+sed -e "s/NoColorVal/$profileColor/" \
+    -e "s/NoDescriptionVal/$profileDescription/" \
+    -e "s/NoMoldTypeVal/$profileMoldType/" \
+    < banana.config.json > cmd-cli.config.json
+exitOnFailure "Failed to update config file." $?
+
+# show the property values that will be used
 cmd-cli profile mapping
-CMDRC=$?
-if [ $CMDRC -gt 0 ]
-then
-    echo "Profile mapping command failed!" 1>&2
-    exit $CMDRC
-fi
+exitOnFailure "The 'profile mapping' command failed." $?