+++ packages/imperative/__tests__/__integration__/cmd/__tests__/integration/cli/profiles/__scripts__/profiles/specify_env_for_positional.sh
@@ -1,9 +1,13 @@
-#!/bin/bash
+#!/bin/sh
 
-CMD_CLI_OPT_COLOR="$1" CMD_CLI_OPT_BANANA_DESCRIPTION="$2" CMD_CLI_OPT_MOLD_TYPE="$3" cmd-cli profile mapping-positional
-CMDRC=$?
-if [ $CMDRC -gt 0 ]
-then
-    echo "Profile mapping command failed!" 1>&2
-    exit $CMDRC
-fi
+envColor=$1
+envDescription=$2
+envMoldType=$3
+
+# include exitOnFailure function
+myScriptDir=`dirname $0`
+. $myScriptDir/exitOnFailure.sh
+
+CMD_CLI_OPT_COLOR="$envColor" CMD_CLI_OPT_BANANA_DESCRIPTION="$envDescription" CMD_CLI_OPT_MOLD_TYPE="$envMoldType" \
+    cmd-cli profile mapping-positional
+exitOnFailure "The 'profile mapping-positional' command failed." $?