+++ packages/imperative/__tests__/__integration__/cmd/__tests__/integration/cli/read/__scripts__/profile/create_and_read.sh
@@ -1,4 +1,13 @@
-#!/bin/bash
-set -e
-OUTPUT=$(cmd-cli profiles create insecure "test_insecure" --info "some info" --secret "not so secret info")
+#!/bin/sh
+
+# include exitOnFailure function
+myScriptDir=`dirname $0`
+. $myScriptDir/exitOnFailure.sh
+
+# copy pre-existing profiles to test directory
+cp -r $myScriptDir/../__resources__/profiles profiles
+exitOnFailure "Failed to copy test profile." $?
+
+# read the profile and display its information
 cmd-cli read profile
+exitOnFailure "Failed display profile." $?