+++ packages/imperative/__tests__/__integration__/cmd/__tests__/integration/cli/root/__scripts__/invalid_command2.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 echo "================ISSUING CMD WITH VALID GROUP AND INVALID CMD==============="
-cmd-cli profiles cre abc
+cmd-cli auth login vegetable
 
 exit $?